Don't use tinyint for group id

+++ database/migrations/2022_05_07_200320_create_teams_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('teams', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedTinyInteger('group');
+            $table->unsignedSmallInteger('group');
             $table->timestamps();
         });
     }
